aws-multi-agent-ops/requirements.txt
boto3>=1.28.0
flask>=3.0.0
pyyaml>=6.0
rich>=13.0.0
jinja2>=3.1.0
